requirements.txt
streamlit>=1.28.0
playwright>=1.40.0
openai>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
cryptography>=41.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
